+++ servlet-jsp/multi-api/src/main/java/multi/api/contract/OrderApi.java
@@ -3,7 +3,6 @@ package multi.api.contract;
 import multi.api.dto.basket.OrderDto;
 import multi.api.dto.order.OrderRequest;
 import multi.api.dto.order.OrderResponse;
-import multi.api.exception.basket.BasketDbException;
 import multi.domain.StatusOrder;
 
 import java.util.List;
