--- servlet-jsp/multi-api/src/main/java/multi/api/exception/basket/BasketDbException.java
@@ -1,7 +1,0 @@
-package multi.api.exception.basket;
-
-public class BasketDbException extends RuntimeException{
-    public BasketDbException(String message) {
-        super(message);
-    }
-}
