--- servlet-jsp/multi-api/src/main/java/multi/api/exception/client/ClientDbException.java
@@ -1,7 +1,0 @@
-package multi.api.exception.client;
-
-public class ClientDbException extends RuntimeException{
-    public ClientDbException(String message) {
-        super(message);
-    }
-}
