--- servlet-jsp/multi-api/src/main/java/multi/api/exception/client/ClientValidatorException.java
@@ -1,7 +1,0 @@
-package multi.api.exception.client;
-
-public class ClientValidatorException extends Exception{
-    public ClientValidatorException(String message) {
-        super(message);
-    }
-}
