--- servlet-jsp/multi-api/src/main/java/multi/api/exception/client/UserNullException.java
@@ -1,7 +1,0 @@
-package multi.api.exception.client;
-
-public class UserNullException extends Exception{
-    public UserNullException(String message) {
-        super(message);
-    }
-}
