--- servlet-jsp/multi-api/src/main/java/multi/api/exception/order/OrderDbException.java
@@ -1,7 +1,0 @@
-package multi.api.exception.order;
-
-public class OrderDbException extends RuntimeException{
-    public OrderDbException(String message) {
-        super(message);
-    }
-}
