--- servlet-jsp/multi-api/src/main/java/multi/api/exception/product/ProductDbException.java
@@ -1,7 +1,0 @@
-package multi.api.exception.product;
-
-public class ProductDbException extends RuntimeException{
-    public ProductDbException(String message) {
-        super(message);
-    }
-}
