--- servlet-jsp/multi-api/src/main/java/multi/api/exception/product/ProductValidatorException.java
@@ -1,7 +1,0 @@
-package multi.api.exception.product;
-
-public class ProductValidatorException extends Exception {
-    public ProductValidatorException(String message) {
-        super(message);
-    }
-}
