+++ servlet-jsp/multi-basic/src/main/java/multi/basic/exception/basket/BasketDbException.java
@@ -1,0 +1,7 @@
+package multi.basic.exception.basket;
+
+public class BasketDbException extends RuntimeException {
+    public BasketDbException(String message) {
+        super(message);
+    }
+}
