+++ servlet-jsp/multi-basic/src/main/java/multi/basic/exception/client/ClientDbException.java
@@ -1,0 +1,7 @@
+package multi.basic.exception.client;
+
+public class ClientDbException extends RuntimeException {
+    public ClientDbException(String message) {
+        super(message);
+    }
+}
