+++ servlet-jsp/multi-basic/src/main/java/multi/basic/exception/client/ClientValidatorException.java
@@ -1,0 +1,7 @@
+package multi.basic.exception.client;
+
+public class ClientValidatorException extends RuntimeException {
+    public ClientValidatorException(String message) {
+        super(message);
+    }
+}
