+++ servlet-jsp/multi-basic/src/main/java/multi/basic/exception/client/UserNullException.java
@@ -1,0 +1,7 @@
+package multi.basic.exception.client;
+
+public class UserNullException extends RuntimeException {
+    public UserNullException(String message) {
+        super(message);
+    }
+}
