+++ servlet-jsp/multi-basic/src/main/java/multi/basic/exception/order/OrderDbException.java
@@ -1,0 +1,7 @@
+package multi.basic.exception.order;
+
+public class OrderDbException extends RuntimeException {
+    public OrderDbException(String message) {
+        super(message);
+    }
+}
