+++ servlet-jsp/multi-basic/src/main/java/multi/basic/exception/product/ProductDbException.java
@@ -1,0 +1,7 @@
+package multi.basic.exception.product;
+
+public class ProductDbException extends RuntimeException {
+    public ProductDbException(String message) {
+        super(message);
+    }
+}
