+++ servlet-jsp/multi-basic/src/main/java/multi/basic/exception/product/ProductValidatorException.java
@@ -1,0 +1,7 @@
+package multi.basic.exception.product;
+
+public class ProductValidatorException extends RuntimeException {
+    public ProductValidatorException(String message) {
+        super(message);
+    }
+}
